chore: classroom에 대한 패키지 관리 변경 사항 반영
