+++ src/main/java/com/idealstudy/mvp/application/service/classroom/preclass/FAQService.java
@@ -1,8 +1,8 @@
-package com.idealstudy.mvp.application.service.classroom;
+package com.idealstudy.mvp.application.service.classroom.preclass;
 
 import com.idealstudy.mvp.application.dto.PageRequestDto;
-import com.idealstudy.mvp.application.dto.classroom.FAQDto;
-import com.idealstudy.mvp.application.dto.classroom.FAQPageResultDto;
+import com.idealstudy.mvp.application.dto.classroom.preclass.FAQDto;
+import com.idealstudy.mvp.application.dto.classroom.preclass.FAQPageResultDto;
 import com.idealstudy.mvp.infrastructure.FAQRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
