+++ src/main/java/com/idealstudy/mvp/infrastructure/jpa/repository/classroom/preclass/ClassInquiryJpaRepository.java
@@ -1,6 +1,6 @@
-package com.idealstudy.mvp.infrastructure.jpa.repository.classroom;
+package com.idealstudy.mvp.infrastructure.jpa.repository.classroom.preclass;
 
-import com.idealstudy.mvp.infrastructure.jpa.entity.classroom.ClassInquiryEntity;
+import com.idealstudy.mvp.infrastructure.jpa.entity.classroom.preclass.ClassInquiryEntity;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
