+++ src/main/java/com/idealstudy/mvp/infrastructure/jpa/repository/classroom/preclass/FAQJpaRepository.java
@@ -1,6 +1,6 @@
-package com.idealstudy.mvp.infrastructure.jpa.repository.classroom;
+package com.idealstudy.mvp.infrastructure.jpa.repository.classroom.preclass;
 
-import com.idealstudy.mvp.infrastructure.jpa.entity.classroom.FAQEntity;
+import com.idealstudy.mvp.infrastructure.jpa.entity.classroom.preclass.FAQEntity;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
